Add hx-delete to the delete contact button.
This is by far my least favorite of the HTMX features. I know that people _love_ their `DELETE` verb; but, to me, it feels like it cuts against the grain of the ColdFusion applicaiton server. I'd prefer not to use it.

+++ v2/delete.cfm
@@ -11,6 +11,17 @@
 
 		try {
 
+			// This is a test to see how the workflow responds to error messages.
+			// Otherwise, we don't have a good way to trigger an error in this workflow.
+			if ( contact.name.reFindNoCase( "^Diva\b" ) ) {
+
+				throw(
+					type = "Model.Validation",
+					message = "Contacts with the name Diva cannot be deleted."
+				);
+
+			}
+
 			contactModel.deleteByFilter( id = contact.id );
 
 			goto( "index.cfm?flash=contact.deleted" );
